llm/agents/openai_agent.py
# agents/openai_agent.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class OpenAIAgent:
    """
    Agent for interacting with OpenAI's API.
    This is a placeholder implementation that will be expanded in Step 2.
    """
    
    def __init__(self):
        self.api_key = os.getenv("OPENAI_API_KEY")
        if not self.api_key:
            raise ValueError("OPENAI_API_KEY environment variable is not set")
    
    def test_connection(self, input_text: str):
        """
        Test the connection to OpenAI API.
        This is a placeholder implementation that will be expanded in Step 2.
        """
        # This is a placeholder that will be implemented in Step 2
        return {
            "status": "placeholder",
            "message": f"OpenAI agent received: {input_text}",
            "note": "This is a placeholder implementation that will be expanded in Step 2."
        }