llm/requirements.txt
fastapi
uvicorn
python-dotenv
pytest
httpx
openai
requests